backend/requirements.txt
# === Core Django & REST ===
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0

# === Static & Production Handling ===
whitenoise==6.7.0
gunicorn==23.0.0

# === Utilities ===
Pillow==11.0.0
python-dotenv==1.0.1

# === Optional (if you plan to switch DB later) ===
# psycopg2-binary==2.9.10   # PostgreSQL
# mysqlclient==2.2.4        # MySQL

# === JWT & Token Blacklist Support ===
PyJWT==2.10.1
